Creating quickstart.py, moving caltrack_daily and caltrack_hourly here to avoid problems of circularity.
