+++ eemeter/__init__.py
@@ -24,7 +24,6 @@ from .__version__ import __title__, __description__, __url__, __version__
 from .__version__ import __author__, __author_email__, __license__
 from .__version__ import __copyright__
 from .caltrack import *
-from .processing import *
 from .derivatives import *
 from .exceptions import *
 from .features import *
