+++ eemeter/caltrack/__init__.py
@@ -20,3 +20,4 @@
 from .design_matrices import *
 from .hourly import *
 from .usage_per_day import *
+from .quickstart import *
